Detect MEGA in UWP
