+++ Files/Filesystem/Cloud/Providers/MegaCloudProvider.cs
@@ -1,11 +1,8 @@
 ﻿using Files.Enums;
-using Files.Helpers;
+using Microsoft.Win32;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
-using Windows.ApplicationModel.AppService;
-using Windows.Foundation.Collections;
 
 namespace Files.Filesystem.Cloud.Providers
 {
@@ -15,38 +12,62 @@ namespace Files.Filesystem.Cloud.Providers
         {
             try
             {
-                var connection = await AppServiceConnectionHelper.Instance;
-                if (connection != null)
-                {
-                    var (status, response) = await connection.SendMessageForResponseAsync(new ValueSet()
-                    {
-                        { "Arguments", "DetectMEGASync" }
-                    });
-                    if (status == AppServiceResponseStatus.Success && response.ContainsKey("Count"))
-                    {
-                        var results = new List<CloudProvider>();
-                        foreach (var key in response.Keys
-                            .Where(k => k != "Count" && k != "RequestID")
-                            .OrderBy(o => o))
-                        {
-                            results.Add(new CloudProvider()
-                            {
-                                ID = CloudProviders.Mega,
-                                Name = $"MEGA ({key})",
-                                SyncFolder = (string)response[key]
-                            });
-                        }
-
-                        return results;
-                    }
-                }
-                return Array.Empty<CloudProvider>();
+                return await Task.Run(() => DetectFromRegistry());
             }
             catch
             {
                 // Not detected
                 return Array.Empty<CloudProvider>();
             }
+        }
+
+        private IList<CloudProvider> DetectFromRegistry()
+        {
+            var results = new List<CloudProvider>();
+            using var clsidKey = Registry.ClassesRoot.OpenSubKey(@"CLSID");
+            foreach (var subKeyName in clsidKey.GetSubKeyNames())
+            {
+                using var subKey = clsidKey.OpenSubKey(subKeyName);
+                if ((int?)subKey.GetValue("System.IsPinnedToNameSpaceTree") == 1)
+                {
+                    using var namespaceKey = Registry.CurrentUser.OpenSubKey($@"SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Desktop\NameSpace\{subKeyName}");
+                    var driveType = (string)namespaceKey?.GetValue("");
+                    if (driveType == null)
+                    {
+                        continue;
+                    }
+
+                    using var bagKey = subKey.OpenSubKey(@"Instance\InitPropertyBag");
+                    var syncedFolder = (string)bagKey?.GetValue("TargetFolderPath");
+                    if (syncedFolder == null)
+                    {
+                        continue;
+                    }
+
+                    // Also works for OneDrive, Box, Amazon Drive, iCloudDrive, Dropbox
+                    CloudProviders? driveID = driveType switch
+                    {
+                        "MEGA" => CloudProviders.Mega,
+                        _ => null
+                    };
+                    if (driveID == null)
+                    {
+                        continue;
+                    }
+
+                    results.Add(new CloudProvider()
+                    {
+                        ID = driveID.Value,
+                        Name = driveID switch
+                        {
+                            CloudProviders.Mega => $"MEGA ({System.IO.Path.GetFileName(syncedFolder.TrimEnd('\\'))})",
+                            _ => null
+                        },
+                        SyncFolder = syncedFolder
+                    });
+                }
+            }
+            return results;
         }
     }
 }
